requirements.txt
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
authlib==1.6.0
bcrypt==4.3.0
black==25.1.0
certifi==2025.4.26
cfgv==3.4.0
click==8.2.0
colorama==0.4.6
distlib==0.3.9
dnspython==2.7.0
ecdsa==0.19.1
elasticsearch==9.0.1
email_validator==2.2.0
fastapi==0.115.12
filelock==3.18.0
greenlet==3.2.2
grep==0.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
identify==2.6.10
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Mako==1.3.10
MarkupSafe==3.0.2
mypy==1.15.0
mypy_extensions==1.1.0
nodeenv==1.9.1
packaging==25.0
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.3.8
pluggy==1.6.0
pre_commit==4.2.0
psycopg2-binary==2.9.10
pyasn1==0.4.8
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
pytest==8.3.5
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
PyYAML==6.0.2
rsa==4.9.1
ruff==0.11.10
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2